+++ lib/service/social.dart
@@ -849,7 +849,7 @@ class Social extends Service with NotificationsListener {
     }
   }
 
-  Future<List<Conversation>?> broadcastIndividualMessage({required ContextItem context, required String message, List<FileAttachment>? fileAttachments, Map<String, dynamic>? extraParams }) async {
+  Future<List<Conversation>?> broadcastIndividualMessage({required ContextItem context, required String message, List<FileAttachment>? fileAttachments, List<String>? recepientIds }) async {
     String? socialUrl = Config().socialUrl;
     if (StringUtils.isEmpty(socialUrl)) {
       Log.e('Failed to broadcast individual message for context ${context}. Reason: missing social url.');
@@ -863,8 +863,10 @@ class Social extends Service with NotificationsListener {
       'context': context.toJson(),
       'message': message,
       'file_attachments': FileAttachment.listToJson(fileAttachments),
-      if (extraParams != null)
-        ...extraParams,
+      if (recepientIds?.isNotEmpty == true)
+        'recipients' : recepientIds,
+      if (recepientIds?.isNotEmpty != true)
+        'all_group_members': true,
     });
     Response? response = await Network().post('$socialUrl/conversations/broadcast-individual', auth: Auth2(), body: requestBody);
     int? responseCode = response?.statusCode;